Rename Error404.vue since it was also used for error 403
