+++ app/client/src/pages/ViewProject.vue
@@ -513,7 +513,7 @@ import Banner from '../components/Banners/Banner'
 import ProgressBar from '../components/ProgressBar'
 import MarkdownTranslator from '../components/MarkdownTranslator'
 
-import NotFound from './Error404'
+import NotFound from './NotFound'
 
 export default {
   components: {
